Fix lexical/syntactic grammar for namespace names
The spec had namespace names and qualified names defined in the lexical
grammar, but no way to get from "token" to "qualified name". The
simplest solution is to make "\" a token and make the namespace name
and qualified name be part of the syntactic grammar, not the lexical
grammar.

I note that this is how the Hack and PHP tokenizers actually work, so
the spec is brought into line with the implementation.