Fix small errors in lexical grammar (#74)
* Fix three typos involving backslash characters

A backslash was somehow used where a tilde was expected in two places,
and an extra backslash crept into an arithmetic expression.

* Add missing coalesce operator to grammar summary

When the coalescing operator (??) was added to section 9, it was not
added to the copy of the grammar in section 22.

* Fix lexical/syntactic grammar for namespace names

The spec had namespace names and qualified names defined in the lexical
grammar, but no way to get from "token" to "qualified name". The
simplest solution is to make "\" a token and make the namespace name
and qualified name be part of the syntactic grammar, not the lexical
grammar.

I note that this is how the Hack and PHP tokenizers actually work, so
the spec is brought into line with the implementation.

* Add ellipses to lexical grammar

The lexical grammar did not include the "..." token used in variadic
formal parameter lists.