Small PHP difference clarifications
* Hack reserves keywords "parent" and "self"; this was not documented
in the specification. (Interestingly, Hack does not reserve "null",
"true", "false" or "this", which we might want to look into.)

* The explanation of the ${ } interpolation in string literals had a
link to a now-deleted section of the specification, which meant that
the feature was not explained. Since this feature is very restricted
compared to its equivalent in PHP -- the PHP version of it cannot even
be *lexed* without direction from the parser! -- I've added some
verbiage explaining the difference and giving examples.

* The deleted feature mentioned in the point above is the Variable-Name
creation operator, but does not give the syntax which has been removed.
The ${$y} syntax is at present unsupported in Hack. The $$y syntax is
supported incorrectly now; Alex has a diff outstanding to make it an
error.