Master (#75)
* Fix three typos involving backslash characters

A backslash was somehow used where a tilde was expected in two places,
and an extra backslash crept into an arithmetic expression.

* Add missing coalesce operator to grammar summary

When the coalescing operator (??) was added to section 9, it was not
added to the copy of the grammar in section 22.

* Fix lexical/syntactic grammar for namespace names

The spec had namespace names and qualified names defined in the lexical
grammar, but no way to get from "token" to "qualified name". The
simplest solution is to make "\" a token and make the namespace name
and qualified name be part of the syntactic grammar, not the lexical
grammar.

I note that this is how the Hack and PHP tokenizers actually work, so
the spec is brought into line with the implementation.

* Add ellipses to lexical grammar

The lexical grammar did not include the "..." token used in variadic
formal parameter lists.

* Small PHP difference clarifications

* Hack reserves keywords "parent" and "self"; this was not documented
in the specification. (Interestingly, Hack does not reserve "null",
"true", "false" or "this", which we might want to look into.)

* The explanation of the ${ } interpolation in string literals had a
link to a now-deleted section of the specification, which meant that
the feature was not explained. Since this feature is very restricted
compared to its equivalent in PHP -- the PHP version of it cannot even
be *lexed* without direction from the parser! -- I've added some
verbiage explaining the difference and giving examples.

* The deleted feature mentioned in the point above is the Variable-Name
creation operator, but does not give the syntax which has been removed.
The ${$y} syntax is at present unsupported in Hack. The $$y syntax is
supported incorrectly now; Alex has a diff outstanding to make it an
error.

* Link to enum-declarations was wrong

Link to enum-declarations was wrong

* Correct require-once-expression to require-once-directive

Looks like a previous search-and-replace operation was incomplete. The
require-once directive should be listed as a directive, not an
expression.

* The ${} interpolation is invalid in strict Hack

The ${} interpolation is invalid in strict Hack, so I've removed it
from the specification entirely. (Except to be called out as
unsupported in section 23.)

* Fixed missing hyphens in return-type

Fixed missing hyphens in return-type

* Fix broken cross references

Fix broken cross references