Namespaces contain declarations, not statements
The grammar suggests that the "namespace foo { }" form contains a block
of statements, but in fact it contains a block of declarations.

I've also eliminated the problematic use of "scope" to mean "definition
extent". We already define "scope" as the region of program text in
which a given name is visible, so we should not use "scope" to mean the
region of program text which belongs syntactically to a namespace
definition.