Fix parameter declaration in catch clause
The text originally said that a catch clause parameter is a parameter
list with exactly one element. But this is wrong; a parameter list with
exactly one element can (1) have an optional attribute, (2) have an
optional default value, and (3) have an optional trailing comma. None
of those are legal in Hack.

It is far simpler to say that a catch clause must have a type and a
variable, end of story.