Fix up type aliases
I could not make head nor tail of the constraints on type aliases; as
far as I can tell, the thing to the right can be any type, and I don't
know why there is all this verbiage -- some of it ungrammatical and
likely due to mis-edits -- about qualified names and what the name
means.

Also, a type alias may declare a generic type parameter list. I'm not
yet quite sure of what the semantics are. Once I figure that out I'll
come back and add some text describing it.