Describe heuristic for disambiguating casts
The spec only allowed for four things in a cast operator, but in fact
Hack allows eight keywords and any simple name. The grammar now matches
the implementation.

However, doing so introduces an ambiguity; (x)-$y could be a
subtraction or a cast.  The spec now describes a heuristic that an
implementation can use to tell whether or not (x) is to be parsed as a
cast or a parenthesized expression.

There is still more work to do here, not done in this diff:

(1) We need to describe the semantics of all possible casts.
(2) We should consider whether the cast type can be more than just a
simple name.  Can it be an array type, a shape type, a generic type, a
nullable type?  Can it be a qualified name?  And so on.